changed ints to decimals for averages and costs

+++ schema.sql
@@ -7,15 +7,15 @@ USE project_nomad_booking;
 
 CREATE TABLE listings (
   id int PRIMARY KEY,
-  avg_cost_per_night int NOT NULL,
+  avg_cost_per_night decimal(10, 2) NOT NULL,
   avg_rating decimal (3, 2),
   max_adults int NOT NULL,
   max_children int NOT NULL,
   max_infants int NOT NULL,
-  cleaning_fee int NOT NULL,
+  cleaning_fee decimal(10,2) NOT NULL,
   service_fee_perc decimal (3, 2) NOT NULL,
   occ_tax_rate_perc decimal (3, 2) NOT NULL,
-  additional_guest_fee int NOT NULL
+  additional_guest_fee decimal(10,2) NOT NULL
 );
 
 CREATE TABLE reservations (
@@ -28,7 +28,7 @@ CREATE TABLE reservations (
 CREATE TABLE listing_daily_costs (
   id int PRIMARY KEY,
   listing_id int NOT NULL,
-  cost_per_night int NOT NULL,
+  cost_per_night decimal(10,2) NOT NULL,
   start_date date NOT NULL,
   end_date date NOT NULL
 );
